Refactor PDF to be be a dynamic resource route
